% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_table.R
\name{df_table}
\alias{df_table}
\title{Experimental: turn \code{data.frame} objects into \code{sc_data}}
\usage{
df_table(x, label = "data")
}
\arguments{
\item{x}{a \code{data.frame} or an object that is compatible with \code{as.data.frame}}

\item{label}{a character vector of length one that will be used as a
display label for the dataset.}
}
\description{
Adapter function to initialize \code{sc_data} objects based on a \code{data.frame}.
Currently, there are heavy restrictions on how the data should look
like to work with this function - see Details. This function is not exported.
}
\details{
Codes will be automatically assigned to each field, measure and level
in the data. Allowed column types are \code{numeric}, \code{integer}, \code{factor}
and \code{character}.

There are several improvements that need to be made before this will be
an exported function
\itemize{
\item \code{sc_tabulate()} currently assumes that each combination of field
values occurs once in the data. This is not necessarily the case here.
\item Time columns are not supported.
}
}
\examples{
STATcubeR:::df_table(iris)
}
\keyword{internal}
