% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statcube_table.R
\name{get_statcube_table}
\alias{get_statcube_table}
\title{Führe eine API Abfrage gegen STATcube durch}
\usage{
get_statcube_table(file = lgr01(), token = statcube_token())
}
\arguments{
\item{file}{Pfad zu einem JSON file, welches über die STATcube GUI
heruntergeladen wurde (Open Data API Abfrage)}

\item{token}{Ein API Token, welcher via
\code{\link[=statcube_browse_preferences]{statcube_browse_preferences()}} angezeigt werden kann.}
}
\value{
Ein Objekt der Klasse \code{STATcube_response} welche den Rückgabewert
von \code{\link[httr:POST]{httr::POST()}} beinhaltet
}
\description{
Erlaube es anhand eines jsone Files eine Tabelle von STATcube abzurufen.
}
\examples{
lgr_01 <- get_statcube_table()
}
