% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_tabulate.R
\name{od_tabulate}
\alias{od_tabulate}
\title{Turn od_table objects into tidy data frames}
\usage{
od_tabulate(
  table,
  ...,
  .list = NULL,
  parse_time = TRUE,
  round = TRUE,
  recode_zeros = TRUE
)
}
\arguments{
\item{table}{An object of class \code{od_table}}

\item{...}{Names of measures and/or fields}

\item{.list}{allows to define the arguments for \code{...} as a character vector.}

\item{parse_time}{should field variables of type time be converted
into a \code{POSIXct} format? (currently not respected)}

\item{round}{use the precision of each measure for rounding (currently not
respected as this meta-information is not available)}

\item{recode_zeros}{turn zero values into \code{NA}s prior to rounding}
}
\description{
\code{od_tabulate()} extracts the data in the table and turns it into a tidy
data.frame. It is basically a copy of \code{\link[=sc_tabulate]{sc_tabulate()}}. Additionally,
certain STATcube conventions are applied
\itemize{
\item zeros are recoded to \code{NA}s
\item rounding is done according to the precision of each measure
}

the \code{...} argument decides which measures and/or fields should be included
in the output. If no measures are given, all measures are included. The same
is true for fields.
}
\examples{
\dontrun{
table <- od_table(od_list()$id[1])

od_tabulate(table)
od_tabulate(table, "Geschlecht", "Staatsangehörigkeit")
}
}
