% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saved_table.R
\name{sc_saved_tables_list}
\alias{sc_saved_tables_list}
\alias{sc_saved_table}
\title{Importiere gespeicherte Tabellen}
\usage{
sc_saved_tables_list()

sc_saved_table(table_uri, token = sc_token())
}
\arguments{
\item{table_uri}{Identifikator einer gespeicherten Tabelle aus \code{\link[=sc_saved_tables_list]{sc_saved_tables_list()}}}

\item{token}{Ein API Token}
}
\description{
Funktionen um gespeicherte Tabellen aus STATcube zu importieren
}
\examples{
\dontrun{

(saved_tables <- sc_saved_tables_list())
table_uri <- saved_tables$id[1]
my_response <- sc_saved_table(table_uri)
as.data.frame(my_response)
}
}
