% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{sc_browse}
\alias{sc_browse}
\alias{sc_browse_preferences}
\alias{sc_browse_ogd}
\title{Links to important STATcube and OGD pages}
\usage{
sc_browse()

sc_browse_preferences()

sc_browse_ogd()
}
\description{
A collection of links, to browse important STATcube pages.
}
\examples{
# Show starting page
sc_browse()

# Show the preferences page with the API key
sc_browse_preferences()

# Show the landing page for OGD datasets
sc_browse_ogd()
}
