% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sc_cache}
\alias{sc_cache}
\alias{sc_cache_enable}
\alias{sc_cache_disable}
\alias{sc_cache_clear}
\title{Cache responses from the /table endpoint}
\usage{
## enable caching for the current R session
sc_cache_enable()

## disable caching for the current R session
sc_cache_disable()

## remove all files from the cache
sc_cache_clear()
}
\description{
Functions to cache all requested tables in the directory \verb{~/.STATcubeR_cache}
and reuse them in calls to \code{sc_table()} and \code{sc_table_custom()}. These
functions are designed for testing and documentation and should not be
regarded as part of the STATcubeR interface. The caching logic is likely to
change in the future in which case \code{sc_cache_clear()} is required to purge
old cache entries.
}
\keyword{internal}
