% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{sc_catalogue}
\alias{sc_catalogue}
\title{Obtain a catalogue of available tables and databases}
\usage{
sc_catalogue(key = sc_key())
}
\arguments{
\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}
}
\description{
Invoke the /schema endpoint to recurse into all datasets and tables. The
return value contains a nested list with ids and labels for all
resources
}
\examples{
my_catalogue <- sc_catalogue()

## print
my_catalogue

## access the parsed catalogue
arbeitsmarkt <- my_catalogue$parsed$Statistiken$Arbeitsmarkt
arbeitsmarkt$`Mikrozensus-Arbeitskräfteerhebung Jahresdaten`
}
