% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_db_info.R
\name{sc_db_info}
\alias{sc_db_info}
\alias{print.sc_db_info}
\title{Get information about a database}
\usage{
sc_db_info(db_id, key = sc_key())

\method{print}{sc_db_info}(x, functions = FALSE, value = FALSE, ...)
}
\arguments{
\item{db_id}{a database id}

\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{x}{object to be printed}

\item{functions}{show resources of type \code{STAT_FUNCTION}?}

\item{value}{show resources of type \code{VALUE}?}

\item{...}{ignored}
}
\description{
Invoke the /schema endpoint and recurse into all valuesets. The return
value can be displayed as a tree object
}
\examples{
my_db_info <- sc_db_info("devgrrgr004")

# printing
my_db_info
print(my_db_info, value = TRUE)

# access parsed table
my_db_info$parsed[2:7, 1:3]

# access the raw response from httr::GET()
my_response <- my_db_info$response
my_response$headers$date
my_content <- httr::content(my_response)
my_content$label
}
