% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{sc_get_response}
\alias{sc_get_response}
\alias{sc_example}
\title{Post an api request against STATcube}
\usage{
sc_get_response(file, token = sc_token())

sc_example(filename)
}
\arguments{
\item{file}{path to a json file, which was downloaded via the STATcube
gui ("Open Data API Abfrage")}

\item{token}{(\code{string}) An API token. To display your token, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{filename}{The name of an example json file.}
}
\value{
An object of class \code{STATcube_response} which contains the return
value of \code{\link[httr:POST]{httr::POST()}}
}
\description{
retrieve a response from STATcube based on a json file
}
\examples{
\dontrun{
lgr_01 <- sc_get_response(sc_example("LGR01.json"))
}
}
