% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_json.R
\name{sc_json_get_server}
\alias{sc_json_get_server}
\title{Get the server from a json request}
\usage{
sc_json_get_server(json)
}
\arguments{
\item{json}{path to a request json}
}
\value{
\code{"ext"}, \code{"red"} or \code{"prod"} depending on the database uri in the
json request
}
\description{
parses a json request and returns a short string representing
the corresponding STATcube server
}
\examples{
sc_json_get_server(sc_example('accomodation'))
}
