% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_meta.R
\name{sc_meta}
\alias{sc_meta}
\alias{sc_meta_field}
\title{Get metadata for a STATcube table}
\usage{
sc_meta(response)

sc_meta_field(response, i = 1)
}
\arguments{
\item{response}{An object of class \code{sc_table}}

\item{i}{index of the field, for which further metadata are desired}
}
\description{
Functions to extract metadata from a \code{sc_table} object.
}
\seealso{
Other functions for /table: 
\code{\link{sc_table}()},
\code{\link{sc_write_json}()}
}
\concept{functions for /table}
\keyword{internal}
