% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_json.R, R/table_as_data_frame.R
\name{sc_post_json}
\alias{sc_post_json}
\alias{sc_example}
\alias{as.data.frame.STATcube_response}
\title{Post an api request against STATcube}
\usage{
sc_post_json(file, key = sc_key(), language = c("en", "de"))

sc_example(filename)

\method{as.data.frame}{STATcube_response}(
  x,
  ...,
  drop_aggregates = TRUE,
  recode_na = TRUE,
  label_vars = TRUE,
  parse_fields = TRUE
)
}
\arguments{
\item{file}{path to a json file, which was downloaded via the STATcube
gui ("Open Data API Abfrage")}

\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{language}{The language to be used for labeling. \code{"en"} or \code{"de"}}

\item{filename}{The name of an example json file.}

\item{x}{an R object of class \code{STATcube_response}}

\item{...}{unused}

\item{drop_aggregates}{remove rows containing aggregates from the table.
The default (\code{TRUE}) makes the table tidy}

\item{recode_na}{Recode cells to \code{NA} if the corresponding annotations have
the value \code{"X"} (cross tabulation not allowed).}

\item{label_vars}{Use labels for column names. Alternatively, codes are
used}

\item{parse_fields}{Parse field columns as \code{character} or \code{POSIXct}
depending on labels? Alternatively, codes are used}
}
\value{
An object of class \code{STATcube_response} which contains the return
value of \code{\link[httr:POST]{httr::POST()}}
}
\description{
retrieve a response from STATcube based on a json file
}
\examples{
\dontrun{
lgr_01 <- sc_post_json(sc_example("LGR01.json"))
}
}
