% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{sc_token}
\alias{sc_token}
\alias{sc_token_set}
\alias{sc_token_get}
\alias{sc_token_prompt}
\title{Manage your API Token}
\usage{
sc_token()

sc_token_set(token, test = TRUE)

sc_token_get()

sc_token_prompt(test = TRUE)
}
\arguments{
\item{token}{(\code{string}) An API token. To display your token, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{test}{Use a test-requst to verify the token?}
}
\value{
All functions return the token (invisibly)
}
\description{
Functions to get/set the STATcube api token and make them available for calls
against the STATcube api.
}
\details{
\itemize{
\item \code{sc_token()} forwards to \code{sc_token_get()} if the token is already present.
Otherwise, \code{sc_token_prompt()} will be invoked.
}

\itemize{
\item \code{sc_token_set()} can be used to pass the token as a parameter (\code{string})
}

\itemize{
\item \code{sc_token_get()} returns the token, if it exists. Otherwise,
an error is thrown.
}

\itemize{
\item \code{sc_token_prompt()} prompts for a token via \code{\link[=readline]{readline()}}
}
}
