% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_json.R
\name{sc_upload_json}
\alias{sc_upload_json}
\title{Lade JSON API Abfrage aus dem Windows System hoch}
\usage{
sc_upload_json()
}
\value{
Ein Objekt der Klasse \code{STATcube_response}
}
\description{
Diese Funktion startet einen Upload-Dialog, welcher es erlaubt API Abfragen
im JSON-Format aus dem Windows-System zu Importieren. Die JSON-Datei wird
hohgeladen und über \code{\link[=sc_get_response]{sc_get_response()}} ausgeführt.
}
\examples{
\dontrun{

my_table <- sc_upload_json()
}
}
