% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_table.R
\name{sdmx_table}
\alias{sdmx_table}
\title{Import data from SDMX}
\usage{
sdmx_table(file)
}
\arguments{
\item{file}{a "sdmx archive" file that was downloaded from STATcube.}
}
\value{
An object of class \code{sc_data}
}
\description{
Function that reads STATcube data from an sdmx archive - a zip file
consisting of \code{structure.xml} with metadata and \code{dataset.xml} for the
values.
}
\keyword{experimental}
