% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{statcube_browse}
\alias{statcube_browse}
\alias{statcube_browse_preferences}
\title{Links zu wichtigen STATcube Seiten}
\usage{
statcube_browse()

statcube_browse_preferences()
}
\description{
Eine Sammlung von links, um relevante STATcube Seiten anzuzeigen
}
\examples{
# Zeige Startseite
statcube_browse()

# Zeige "Einstellungen"
statcube_browse_preferences()
}
