% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{statcube_token}
\alias{statcube_token}
\title{API Token hinzufügen}
\usage{
statcube_token(token = NULL)
}
\arguments{
\item{token}{(Optional) Ein API Token, welcher via "Benutzerkonto" einsehbar
ist.}
}
\value{
Es wird der API Token (invisible) zurückgegeben
}
\description{
Diese Funktion erlaubt es, einen STATcube API Token in den authSTAT vault zu
speichern bzw den Token aus dem Vault zu laden.
}
