% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_json.R
\name{upload_json}
\alias{upload_json}
\title{Lade json APi Abfrage aus dem Windows system hoch}
\usage{
upload_json()
}
\value{
Ein Objekt der Klasse \code{STATcube_response}
}
\description{
Diese Funktion startet einen Upload-Dialog, welcher es erlaubt API Abfragen
im JSON-Format aus dem Windows-System zu Importieren. Die JSON-Datei wird
hohgeladen und über \code{\link[=get_statcube_response]{get_statcube_response()}} ausgeführt.
}
\examples{
\dontrun{

my_table <- upload_json()
}
}
