% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_list.R
\name{od_catalogue}
\alias{od_catalogue}
\title{Get a catalogue for OGD datasets}
\usage{
od_catalogue(server = "ext", local = TRUE)
}
\arguments{
\item{server}{the OGD-server to be used. \code{"ext"} (the default) for the
external server or \code{prod} for the production server}

\item{local}{If \code{TRUE} (the default), the catalogue is created based on
cached json metadata. Otherwise, the cache is updated prior to
creating the catalogue using a "bulk-download" for metadata files.}
}
\description{
\strong{EXPERIMENTAL} This function parses several json metadata files at once
and combines them into a \code{data.frame} so the datasets can easily be
filtered based on categorizations, tags, number of classifications, etc.
}
\details{
The naming, ordering and choice of the columns is likely to change.
Currently, the following columns are provided
\itemize{
\item \strong{\code{title}} (chr) the title of the dataset
\item \strong{\code{measures}} (int) number of measure variables
\item \strong{\code{fields}} (int) number of classification fields
\item \strong{\code{modified}} (dttm) the timestamp when the dataset was last modified
\item \strong{\code{created}} (dttm) the timestamp when the dataset was created
\item \strong{\code{id}} (chr) the OGD identifier
\item \strong{\code{database}} (chr) the identifier of the corresponding STATcube database
\item \strong{\code{notes}} (chr) a description for the dataset
\item \strong{\code{update_frequency}} (chr) how often is the dataset updated?
\item \strong{\code{tags}} (list) tags assigned
\item \strong{\code{categorization}} (chr) the category of the dataset
\item \strong{\code{json}} (list) the full json metadata
}
}
\examples{
catalogue <- od_catalogue()
catalogue
catalogue$update_frequency \%>\% table()
catalogue$categorization \%>\% table()
catalogue[catalogue$categorization == "Gesundheit", 1:4]
catalogue[catalogue$measures >= 70, 1:3]
catalogue$json[[1]]
catalogue$database \%>\% head()
}
