% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_list.R
\name{od_list}
\alias{od_list}
\title{List available Opendata datasets}
\usage{
od_list(unique = TRUE)
}
\arguments{
\item{unique}{some datasets are pulbished under multiple groups.
They will only be listed once with the first group they appear in unless
this parameter is set to \code{FALSE}.}
}
\value{
a \code{data.frame} with two columns
\itemize{
\item \code{"category"}: Grouping under which a dataset is listed
\item \code{"id"}: Name of the dataset which can later be used in
\code{\link[=od_table]{od_table()}}
\item \code{"label"}: Description of the dataset
}
}
\description{
\code{\link[=od_list]{od_list()}} returns a \code{data.frame } containing all datasets published at
\href{https://data.statistik.gv.at}{data.statistik.gv.at}
}
\examples{
df <- od_list()
df
subset(df, category == "Bildung und Forschung")
}
