% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_list.R
\name{od_list}
\alias{od_list}
\title{Liste alle Datensätze}
\usage{
od_list()
}
\value{
ein \code{data.frame} mit zwei Spalten
\itemize{
\item \code{"gruppe"}: Inhaltliche Gliederungsgruppe, unter der ein Datensatz gegliedert ist.
\item \code{"id"}: Name des Datensatzes; Kann in \code{\link[=od_table]{od_table()}} verwendet werden
\item \code{"label"} Bezeichnung der Datenbank
}
}
\description{
\code{\link[=od_list]{od_list()}} gibt alle auf \href{https://data.statistik.gv.at}{data.statistik.gv.at} verfügbaren Datensätze
zurück.
}
\details{
manche Datensätze sind unter mehreren Gruppen gelistet.
}
\examples{
df <- od_list()
subset(df, df$grupp == "Arbeit")
}
