% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_list.R
\name{od_list}
\alias{od_list}
\title{List available Opendata datasets}
\usage{
od_list()
}
\value{
a \code{data.frame} with two columns
\itemize{
\item \code{"gruppe"}: Grouping under which a dataset is listed
\item \code{"id"}: Name of the dataset which can later be used in
\code{\link[=od_table]{od_table()}}
\item \code{"label"}: Description of the dataset
}
}
\description{
\code{\link[=od_list]{od_list()}} returns a \code{data.frame } containing all datasets published at
\href{https://data.statistik.gv.at}{data.statistik.gv.at}
}
\details{
some datasets are pulbished under multiple groups
}
\examples{
df <- od_list()
subset(df, df$gruppe == "Arbeit")
}
