% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_table.R
\name{od_table}
\alias{od_table}
\title{Create a table-instance for a open-data Dataset (data.statistik.gv.at)}
\usage{
od_table(id)
}
\arguments{
\item{id}{the name of the data-set that should be accessed}
}
\value{
An object of class \code{od_table} which contains the return
value of the \code{\link[httr:POST]{httr::POST()}} request in \code{obj$response}. The object also
provides member functions to parse this response object.
}
\description{
\code{\link[=od_table]{od_table()}} returns an \code{R6}-class object containing all relevant data
and metadata information to be used as input for graphSTAT.
}
\examples{
r1 <- od_table(id = "OGD_veste309_Veste309_1")

r1$meta
r1$field()
r1$field(1)
r1$data

r2 <- od_table(id = "OGD_krebs_ext_KREBS_1")
r2$meta
r2$field()
r2$field(1)
r2$data

r3 <- od_table(id = "OGD_konjunkturmonitor_KonMon_1")
r3$meta
r3$field()
r3$field(1)
r3$data
}
