% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_table.R
\name{od_table}
\alias{od_table}
\title{Create a table-instance from an open-data dataset}
\usage{
od_table(id, language = c("en", "de"))
}
\arguments{
\item{id}{the id of the data-set that should be accessed}

\item{language}{language to be used for labeling. \code{"en"} or \code{"de"}}
}
\value{
The returned objects is of class \code{sc_table} and inherits several parsing
methods from \link{sc_data}. See \link{od_table_class} for the full class
documentation.
}
\description{
\code{od_table(id)} returns an \code{R6}-class object containing all relevant data
and metadata from either https://data.statistik.gv.at/data/ or a locally
generated json-file (using \code{STATgraphUtils::od_generator()}. In the later
case it is assumed that all related files are located in the same folder.
}
\section{Components}{
\tabular{lll}{
   \strong{Component} \tab \strong{Corresponding File on Server} \tab \strong{Type} \cr
   \verb{$data         } \tab \verb{https://data.statistik.gv.at/data/$\{id\}.csv} \tab \code{remote} \cr
   \verb{$header       } \tab \verb{https://data.statistik.gv.at/data/$\{id\}_HEADER.csv} \tab \code{remote} \cr
   \verb{$field(code)  } \tab \verb{https://data.statistik.gv.at/data/$\{id\}_$\{code\}.csv} \tab \code{remote} \cr
   \verb{$json         } \tab \verb{https://data.statistik.gv.at/ogd/json?dataset=$\{id\}} \tab \code{remote} \cr
   \verb{$data         } \tab \verb{/path/to/\{id\}.csv} \tab \code{local} \cr
   \verb{$header       } \tab \verb{/path/to/\{id\}_HEADER.csv} \tab \code{local} \cr
   \verb{$field(code)  } \tab \verb{/path/to/\{id\}_$\{code\}.csv} \tab \code{local} \cr
   \verb{$json         } \tab \verb{/path/to/\{id\}.json} \tab \code{local} \cr
}
}

\examples{
x <- od_table("OGD_krebs_ext_KREBS_1")

## metadata
x
x$meta
x$field("Sex")
x$field(3)

## data
x$data
x$tabulate()

## tabulation: see `?sc_tabulate` for more examples
x$tabulate("Reporting year", "Sex")

## switch language
x$language <- "de"
x
x$tabulate()

## other interesting tables
od_table("OGD_veste309_Veste309_1")
od_table("OGD_konjunkturmonitor_KonMon_1")
od_table("OGD_krankenbewegungen_ex_LEISTUNGEN_1")
od_table("OGD_f1741_HH_Proj_1")
od_table("OGD_veste303_Veste203_1")
}
