% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_table_save.R
\name{od_table_save}
\alias{od_table_save}
\alias{od_table_local}
\title{Saves/load opendata datasets via zip archives}
\usage{
od_table_save(x, file = NULL)

od_table_local(file)
}
\arguments{
\item{x}{an object of class \code{od_table}}

\item{file}{A zip file for the dataset. For \code{od_table_save()}, the default
is \verb{\{id\}.zip} where \code{id} denotes the OGD identifier.}
}
\description{
\code{od_table_save()} creates a zip file containing all relevant data from the
OGD portal. \code{od_table_local()} creates an \code{od_table} object based on
a zipped version of a dataset.
}
\examples{
x <- od_table("OGD_krebs_ext_KREBS_1")

# save dataset to ./OGD_krebs_ext_KREBS_1.zip

od_table_save(x)
od_table_local("OGD_krebs_ext_KREBS_1.zip")

# save dataset to a custom path

od_table_save(x, "~/dataset.zip")
od_table_local("~/dataset.zip")
}
