% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_endpoints.R
\name{other_endpoints}
\alias{other_endpoints}
\alias{sc_info}
\alias{sc_rate_limit}
\title{Other endpoints of the STATcube REST API}
\usage{
sc_info(language = c("en", "de"), key = sc_key())

sc_rate_limit(language = c("en", "de"), key = sc_key())
}
\arguments{
\item{language}{The language to be used for labeling. \code{"en"} or \code{"de"}}

\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}
}
\description{
Utilize the simple endpoints \verb{/info} and \verb{/ratelimit}. Those provide
information about available locales and the amount of requests available
for calls against the \verb{/table} endpoint.
}
\examples{
sc_info()
sc_rate_limit()
}
