% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_saved.R
\name{sc_saved_tables_list}
\alias{sc_saved_tables_list}
\alias{sc_saved_table}
\title{Import saved tables}
\usage{
sc_saved_tables_list()

sc_saved_table(table_uri, key = sc_key(), language = c("en", "de"))
}
\arguments{
\item{table_uri}{Identifier of a saved table as returned by
\code{\link[=sc_saved_tables_list]{sc_saved_tables_list()}}}

\item{key}{An API key}

\item{language}{The language to be used for labeling. \code{"en"} or \code{"de"}}
}
\description{
Functions for importing saved tables from STATcube
}
\examples{
\dontrun{

(saved_tables <- sc_saved_tables_list())
table_uri <- saved_tables$id[1]
my_response <- sc_saved_table(table_uri)
as.data.frame(my_response)
}
}
