% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{sc_browse}
\alias{sc_browse}
\alias{sc_browse_preferences}
\alias{sc_browse_ogd}
\title{Links to important STATcube and OGD pages}
\usage{
sc_browse(server = "ext")

sc_browse_preferences(server = "ext")

sc_browse_ogd()
}
\arguments{
\item{server}{A STATcube API server. Defaults to the external Server via
\code{"ext"}. Oher options are \code{"red"} for the editing server and \code{"prod"} for
the production server.}
}
\description{
A collection of links, to browse important STATcube pages.
}
\examples{
# Show starting page
sc_browse()
# Show the preferences page with the API key
sc_browse_preferences()

# Show the landing page for OGD datasets
sc_browse_ogd()
}
