% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{sc_browse}
\alias{sc_browse}
\alias{sc_browse_preferences}
\title{Links zu wichtigen STATcube Seiten}
\usage{
sc_browse()

sc_browse_preferences()
}
\description{
Eine Sammlung von links, um relevante STATcube Seiten anzuzeigen
}
\examples{
# Zeige Startseite
sc_browse()

# Zeige "Einstellungen"
sc_browse_preferences()
}
