% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_db_info.R
\name{sc_db_info}
\alias{sc_db_info}
\alias{print.sc_schema}
\title{Get information about a database}
\usage{
sc_db_info(db_id, key = sc_key())

\method{print}{sc_schema}(x, limit = 30, value = FALSE, ...)
}
\arguments{
\item{db_id}{a database id}

\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{x}{object to be printed}

\item{limit}{maximum number of entries to be printed}

\item{value}{show resources of type \code{VALUE}?}

\item{...}{ignored}
}
\description{
Invoke the /schema endpoint and recurse into all valuesets. The return
value can be displayed as a tree object
}
\examples{
my_db_info <- sc_db_info("deake005")

# printing
my_db_info

# access child nodes
my_db_info$`Demographic Characteristics`
my_db_info$`Demographic Characteristics`$Gender$Gender
my_db_info$`Demographic Characteristics`$Gender$Gender$male

# access the raw response from httr::GET()
my_response <- attr(my_db_info, "response")
my_response$headers$date
my_content <- httr::content(my_response)
my_content$label
}
