% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{sc_get_response}
\alias{sc_get_response}
\alias{sc_example}
\title{Führe eine API Abfrage gegen STATcube durch}
\usage{
sc_get_response(file, token = sc_token())

sc_example(filename)
}
\arguments{
\item{file}{Pfad zu einem JSON file, welches über die STATcube GUI
heruntergeladen wurde (Open Data API Abfrage)}

\item{token}{(\code{string}) Ein API Token. Ein persönlicher Token kann via
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}} angezeigt werden.}

\item{filename}{Name eines Beispiel-Json files.}
}
\value{
Ein Objekt der Klasse \code{STATcube_response} welche den Rückgabewert
von \code{\link[httr:POST]{httr::POST()}} beinhaltet
}
\description{
Erlaube es anhand eines jsone Files eine Tabelle von STATcube abzurufen.
}
\examples{
\dontrun{
lgr_01 <- sc_get_response(sc_example("LGR01.json"))
}
}
