% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{sc_key}
\alias{sc_key}
\alias{sc_key_set}
\alias{sc_key_get}
\alias{sc_key_prompt}
\title{Manage your API Key}
\usage{
sc_key()

sc_key_set(key, test = TRUE)

sc_key_get()

sc_key_prompt(test = TRUE)
}
\arguments{
\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{test}{Use a test-requst to verify the key?}
}
\value{
All functions return the key (invisibly)
}
\description{
Functions to get/set the STATcube API key and make them available for calls
against the STATcube API.
}
\details{
\itemize{
\item \code{sc_key()} forwards to \code{sc_key_get()} if the key is already present.
Otherwise, \code{sc_key_prompt()} will be invoked.
}

\itemize{
\item \code{sc_key_set()} can be used to pass the key as a parameter (\code{string})
}

\itemize{
\item \code{sc_key_get()} returns the key, if it exists. Otherwise,
an error is thrown.
}

\itemize{
\item \code{sc_key_prompt()} prompts for a key via \code{\link[=readline]{readline()}}
}
}
