% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{sc_meta}
\alias{sc_meta}
\alias{sc_meta_field}
\title{Erhalte Metadaten aus einer STATcube Response}
\usage{
sc_meta(response)

sc_meta_field(response, i)
}
\arguments{
\item{response}{Ein Objekt der Klasse \code{STATcube_response}}

\item{i}{index des Feldes, über das genauere Informationen
gewünscht sind}
}
\description{
Funktionen zum extrahieren von Metadaten
}
\examples{
\dontrun{

my_response <- sc_example("bev_seit_1982") \%>\%
  sc_get_response()
sc_meta(my_response)
sc_meta_field(my_response, 1)
}
}
