% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{sc_meta}
\alias{sc_meta}
\alias{sc_meta_field}
\title{Get metadata for a STATcube table}
\usage{
sc_meta(response)

sc_meta_field(response, i = 1)
}
\arguments{
\item{response}{An object of class \code{STATcube_response}}

\item{i}{index of the field, for which further metadata are desired}
}
\description{
Functions to extract metadata from a \code{STATcube_response} object.
}
\examples{
\dontrun{

my_response <- sc_example("bev_seit_1982") \%>\%
  sc_post_json()
sc_meta(my_response)
sc_meta_field(my_response, 1)
}
}
