% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response.R
\name{sc_post_json}
\alias{sc_post_json}
\alias{sc_example}
\title{Post an api request against STATcube}
\usage{
sc_post_json(file, key = sc_key(), language = c("en", "de"))

sc_example(filename)
}
\arguments{
\item{file}{path to a json file, which was downloaded via the STATcube
gui ("Open Data API Abfrage")}

\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{language}{The language to be used for labeling. \code{"en"} or \code{"de"}}

\item{filename}{The name of an example json file.}
}
\value{
An object of class \code{STATcube_response} which contains the return
value of \code{\link[httr:POST]{httr::POST()}}
}
\description{
retrieve a response from STATcube based on a json file
}
\examples{
\dontrun{
lgr_01 <- sc_post_json(sc_example("LGR01.json"))
}
}
