% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoder.R
\name{sc_recoder}
\alias{sc_recoder}
\title{Recode sc_table objects}
\description{
A collection of methods that can be used to modify an object of class
sc_table by reference. Typical usage is to acces the \code{recode} binding
of an \code{sc_table} object and then use method chaining to perform recode
operations.\preformatted{x <- od_table("OGD_krebs_ext_KREBS_1")
x$recode$
  label_field("C-BERJ-0", "de", "JAHR")$
  label_measure("F-KRE", "de", "Anzahl")
}

See the example section for more details.
}
\examples{
x <- od_table("OGD_krebs_ext_KREBS_1")

x$recode$
  label_field("C-KRE_GESCHLECHT-0", "en", "SEX")$
  label_measure("F-KRE", "en", "NUMBER")$
  level("C-KRE_GESCHLECHT-0", "GESCHLECHT-1", "en", "MALE")

x$language <- "en"
x$tabulate("C-KRE_GESCHLECHT-0", "F-KRE")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sc_recoder$new()}}
\item \href{#method-label_field}{\code{sc_recoder$label_field()}}
\item \href{#method-label_measure}{\code{sc_recoder$label_measure()}}
\item \href{#method-level}{\code{sc_recoder$level()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new recoder instance. This will automatically
be performed during the setup of \code{sc_data} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sc_recoder$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the private environment of an \code{sc_data} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label_field"></a>}}
\if{latex}{\out{\hypertarget{method-label_field}{}}}
\subsection{Method \code{label_field()}}{
Change the label of a field variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sc_recoder$label_field(field, language, new)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{a field code}

\item{\code{language}}{a language, "de" or "en"}

\item{\code{new}}{the new label}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-label_measure"></a>}}
\if{latex}{\out{\hypertarget{method-label_measure}{}}}
\subsection{Method \code{label_measure()}}{
Change the label of a measure variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sc_recoder$label_measure(measure, language, new)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measure}}{a measure code}

\item{\code{language}}{a language "de" or "en"}

\item{\code{new}}{the new label}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-level"></a>}}
\if{latex}{\out{\hypertarget{method-level}{}}}
\subsection{Method \code{level()}}{
Change the labels of a level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sc_recoder$level(field, level, language, new)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{a field code}

\item{\code{level}}{a level code for the field}

\item{\code{language}}{a language "de" or "en"}

\item{\code{new}}{the new label for the level}
}
\if{html}{\out{</div>}}
}
}
}
