% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R, R/schema_catalogue.R, R/schema_db.R
\name{sc_schema}
\alias{sc_schema}
\alias{print.sc_schema}
\alias{sc_schema_catalogue}
\alias{sc_schema_db}
\title{Create a request against the /schema endpoint}
\usage{
sc_schema(
  resource_id = NULL,
  depth = NULL,
  language = c("en", "de"),
  key = sc_key()
)

\method{print}{sc_schema}(x, limit = 30, value = FALSE, ...)

sc_schema_catalogue(depth = "folder", key = sc_key())

sc_schema_db(db_id, depth = "valueset", key = sc_key())
}
\arguments{
\item{resource_id}{A resource identifier un uid format}

\item{depth}{If provided, the request will recurse into the given level.
For datasets, available options are \code{NULL} (no recursion), \code{"folder"},
\code{"field"} and \code{"valueset"}. For the catalogue, only \code{NULL} and \code{"folder"}
are applicable.}

\item{language}{The language to be used for labeling. \code{"en"} or \code{"de"}}

\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{x}{object to be printed}

\item{limit}{maximum number of entries to be printed}

\item{value}{show resources of type \code{VALUE}?}

\item{...}{ignored}

\item{db_id}{a database id}
}
\description{
Invoke the /schema endpoint of the STATcube REST API. In case of
\code{sc_schema_catalogue()}, recurse into all datasets and tables and return a
nested list with ids and labels for all resources. For \code{sc_schema_db()},
recurse into all valuesets and return a list of all resources available
tor the specific dataset. The return values can be displayed as a
tree object.
}
\examples{
\dontrun{

my_catalogue <- sc_schema_catalogue()

## print
my_catalogue

## access the parsed catalogue
my_catalogue$Statistics$`Labour Market`
my_catalogue$Statistics$`Labour Market`$`Working hours (Labour Force Survey)`
}
\dontrun{

db_schema <- sc_schema_db("deake005")

# printing
db_schema

# access child nodes
db_schema$`Demographic Characteristics`
db_schema$`Demographic Characteristics`$Gender$Gender
db_schema$`Demographic Characteristics`$Gender$Gender$male

# access the raw response from httr::GET()
my_response <- attr(db_schema, "response")
my_response$headers$date
my_content <- httr::content(my_response)
my_content$label
}
}
\concept{functions for /schema}
