% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R, R/table_as_data_frame.R
\name{sc_table}
\alias{sc_table}
\alias{sc_example}
\alias{as.data.frame.sc_table}
\title{Create a request against the /table endpoint}
\usage{
sc_table(json_file, language = c("en", "de"), key = sc_key())

sc_example(filename)

\method{as.data.frame}{sc_table}(
  x,
  ...,
  drop_aggregates = TRUE,
  recode_na = TRUE,
  label_vars = TRUE,
  parse_fields = TRUE
)
}
\arguments{
\item{json_file}{path to a json file, which was downloaded via the STATcube
gui ("Open Data API Abfrage")}

\item{language}{The language to be used for labeling. \code{"en"} or \code{"de"}}

\item{key}{(\code{string}) An API key. To display your key, call
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}}.}

\item{filename}{The name of an example json file.}

\item{x}{an R object of class \code{sc_table}}

\item{...}{unused}

\item{drop_aggregates}{remove rows containing aggregates from the table.
The default (\code{TRUE}) makes the table tidy}

\item{recode_na}{Recode cells to \code{NA} if the corresponding annotations have
the value \code{"X"} (cross tabulation not allowed).}

\item{label_vars}{Use labels for column names. Alternatively, codes are
used}

\item{parse_fields}{Parse field columns as \code{character} or \code{POSIXct}
depending on labels? Alternatively, codes are used}
}
\value{
An object of class \code{sc_table} which contains the return
value of \code{\link[httr:POST]{httr::POST()}}
}
\description{
retrieve a response from STATcube based on a json file
}
\examples{
\dontrun{
my_table <- sc_table_json(sc_example("LGR01.json"))

# print
my_table

# get matadata for the table
my_table$meta

# get a data.frame
as.data.frame(my_table)

# get metadata for field 1
my_table$field(1)
}
}
