% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_custom.R
\name{sc_table_custom}
\alias{sc_table_custom}
\title{Create custom tables}
\usage{
sc_table_custom(
  db,
  measures = c(),
  dimensions = c(),
  language = c("en", "de"),
  add_totals = TRUE,
  key = NULL
)
}
\arguments{
\item{db}{The uid of a database}

\item{measures}{A character vector of uids for measures. Can be either of
type \code{MEASURE} or of type \code{STAT_FUNCTION}}

\item{dimensions}{A character vector of dimensions for the cube. Can be
either of type \code{FIELD} or type \code{VALUESET}. Those entries are referred to
as \code{fields} in the parsed API response}
}
\description{
Define requests against the /table endpoint by providing
URIs to databases, measures and fields.
The URIs can be obtained using \code{\link[=sc_schema_db]{sc_schema_db()}}.
See the \href{https://statistikat.github.io/STATcubeR/articles/sc_table_custom.html}{Custom tables article}
for more details.
}
\examples{
sc_table_custom(
  db = "str:database:detouextregsai",
  measures = c(
    "str:statfn:detouextregsai:F-DATA1:F-ANK:SUM",
    "str:measure:detouextregsai:F-DATA1:F-UEB"
  ),
  dimensions = c(
    "str:field:detouextregsai:F-DATA1:C-SDB_TIT-0",
    "str:valueset:detouextregsai:F-DATA1:C-C93-2:C-C93SUM-0"
  )
)
}
\keyword{internal}
