% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{sc_token}
\alias{sc_token}
\alias{sc_token_set}
\alias{sc_token_get}
\alias{sc_token_prompt}
\title{API Token hinzufügen}
\usage{
sc_token()

sc_token_set(token, test = TRUE)

sc_token_get()

sc_token_prompt()
}
\arguments{
\item{token}{(\code{string}) Ein API Token. Ein persönlicher Token kann via
\code{\link[=sc_browse_preferences]{sc_browse_preferences()}} angezeigt werden.}

\item{test}{Soll eine Testabfrage mit dem Token durchgeführt werden?}
}
\value{
Alle vier Funktionen geben den Token (invisible) zurück
}
\description{
Diese Funktion erlaubt es, einen STATcube API Token in den authSTAT vault zu
speichern bzw den Token aus dem Vault zu laden.
}
\details{
\itemize{
\item \code{sc_token()} führt \code{sc_token_get()} oder \code{sc_token_prompt()} aus, je
nachdem, ob der Token verfügbar ist.
}

\itemize{
\item \code{sc_token_set()} kann verwendet werden um den Token als Parameter
(\code{string}) zu übergeben
}

\itemize{
\item \code{sc_token_get()} gibt den Token zurück, falls er existiert. Sonst
wird ein Fehler geworfen.
}

\itemize{
\item \code{sc_token_prompt()} fragt den Token via \code{\link[=readline]{readline()}} ab
}
}
