% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{statcube_token}
\alias{statcube_token}
\alias{statcube_token_set}
\alias{statcube_token_get}
\alias{statcube_token_prompt}
\title{API Token hinzuf체gen}
\usage{
statcube_token()

statcube_token_set(token)

statcube_token_get()

statcube_token_prompt()
}
\arguments{
\item{token}{Ein API Token, welcher via
\code{\link[=statcube_browse_preferences]{statcube_browse_preferences()}} angezeigt werden kann.}
}
\value{
Es wird der API Token (invisible) zur체ckgegeben
}
\description{
Diese Funktion erlaubt es, einen STATcube API Token in den authSTAT vault zu
speichern bzw den Token aus dem Vault zu laden.
}
\details{
\code{statcube_token()} verh채lt sich wie \code{statcube_token_get()}, f체hrt aber
\code{statcube_token_prompt()} aus, falls der Token fehlt.
}
