% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{statcube_token}
\alias{statcube_token}
\alias{statcube_token_set}
\alias{statcube_token_get}
\alias{statcube_token_prompt}
\title{API Token hinzufügen}
\usage{
statcube_token()

statcube_token_set(token)

statcube_token_get()

statcube_token_prompt()
}
\arguments{
\item{token}{(\code{string}) Ein API Token. Ein persönlicher Token kann via
\code{\link[=statcube_browse_preferences]{statcube_browse_preferences()}} angezeigt werden.}
}
\value{
Alle vier Funktionen geben den Token (invisible) zurück
}
\description{
Diese Funktion erlaubt es, einen STATcube API Token in den authSTAT vault zu
speichern bzw den Token aus dem Vault zu laden.
}
\details{
\itemize{
\item \code{statcube_token()} führt \code{statcube_token_get()} oder
\code{statcube_token_prompt()} aus, je nachdem, ob der Token verfügbar ist.
}

\itemize{
\item \code{statcube_token_set()} kann verwendet werden um den Token als Parameter
(\code{string}) zu übergeben
}

\itemize{
\item \code{statcube_token_get()} gibt den Token zurück, falls er existiert. Sonst
wird ein Fehler geworfen.
}

\itemize{
\item \code{statcube_token_prompt()} fragt den Token via \code{\link[=readline]{readline()}} ab
}
}
