% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sc_content}
\alias{sc_content}
\alias{sc_write_json}
\alias{sc_annotation_legend}
\title{Hilfsfunktionen}
\usage{
sc_content(response)

sc_write_json(response, file)

sc_annotation_legend(response)
}
\arguments{
\item{response}{Ein Objekt der Klasse STATcube_response}

\item{file}{Ein Dateiname, unter welchem das json abgespeichert
werden soll.}
}
\description{
Unterstützende Funktionen in STATcubeR
}
