% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sc_content}
\alias{sc_content}
\alias{sc_write_json}
\alias{sc_annotation_legend}
\title{Utility Functions}
\usage{
sc_content(response)

sc_write_json(response, file)

sc_annotation_legend(response)
}
\arguments{
\item{response}{An objekt of class \code{STATcube_response}}

\item{file}{A filename, under which the json should be saved.}
}
\description{
\itemize{
\item \code{sc_content()} returns the raw api response as a nested
list
}

\itemize{
\item \code{sc_write_json()} saves the api request for a table as
a json file. The resulting json file can be passed to \code{\link[=sc_post_json]{sc_post_json()}}
}

\itemize{
\item \code{sc_annotation_legend()} lists all annotations occurring in
a table together with descriptions of the annotations.
}
}
\details{
Supporting functions for STATcubeR
}
