% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sc_write_json}
\alias{sc_write_json}
\alias{sc_annotation_legend}
\title{Utility Functions}
\usage{
sc_write_json(response, file)

sc_annotation_legend(response)
}
\arguments{
\item{response}{an object of class \code{sc_table}}

\item{file}{A filename, under which the json should be saved.}
}
\description{
\itemize{
\item \code{sc_write_json()} saves the api request for a table as
a json file. The resulting json file can be passed to \code{\link[=sc_table]{sc_table()}}
}

\itemize{
\item \code{sc_annotation_legend()} lists all annotations occurring in
a table together with descriptions of the annotations.
}
}
\details{
Supporting functions for STATcubeR
}
